Split out API tests into separate files
These API tests are going to expand quickly, especially
given the fact that we are checking JSON schema definitions.
To keep the API test file from getting massive, this breaks
them out into separate files based on their root path,
this should hopefully make these tests somewhat easier to
manage.
